Moved ObjectSyncInfo to its own file. Tidied up CKRecordConvertible file.
